requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

streamlit==1.2.0
numpy==1.21.4
pandas==1.3.5
opencv-python==4.5.4.60
matplotlib==3.5.0
scikit-image==0.19.0
plotly==5.4.0
tqdm
scipy==1.7.3
pillow==8.4.0
# seaborn==0.11.2
# tensorflow
# tensorflow-addons 
# iterative-stratification
torch==1.10.1+cpu 
torchvision==0.11.1+cpu
# torchaudio==0.10.0+cpu
# graphviz==0.18.2
scikit-learn==1.0.1
# category_encoders
# fastprogress
pretrainedmodels
streamlit-lottie
